+++ package/Runtime/ResourceProviders/Experimental/RemoteTextureProvider.cs
@@ -13,23 +13,40 @@ namespace UnityEngine.ResourceManagement
         internal class InternalOp<TObject> : InternalProviderOperation<TObject>
             where TObject : class
         {
+            IAsyncOperation m_dependencyOperation;
+            AsyncOperation m_requestOperation;
             public InternalProviderOperation<TObject> Start(IResourceLocation location, IAsyncOperation<IList<object>> loadDependencyOperation)
             {
                 m_result = null;
+                m_dependencyOperation = loadDependencyOperation;
+                m_requestOperation = null;
                 if (loadDependencyOperation != null && location != null)
                 {
                     loadDependencyOperation.Completed += (obj) =>
                     {
-                        var reqOp = UnityWebRequestTexture.GetTexture(location.InternalId).SendWebRequest();
-                        if (reqOp.isDone)
-                            DelayedActionManager.AddAction((System.Action<AsyncOperation>)OnComplete, 0, reqOp);
+                        m_requestOperation = UnityWebRequestTexture.GetTexture(location.InternalId).SendWebRequest();
+                        if (m_requestOperation.isDone)
+                            DelayedActionManager.AddAction((System.Action<AsyncOperation>)OnComplete, 0, m_requestOperation);
                         else
-                            reqOp.completed += OnComplete;
+                            m_requestOperation.completed += OnComplete;
                     };
                 }
                 return base.Start(location);
             }
+            public override float PercentComplete
+            {
+                get
+                {
+                    if (IsDone)
+                        return 1;
 
+                    float reqPer = m_requestOperation == null ? 0 : m_requestOperation.progress;
+                    if (m_dependencyOperation == null)
+                        return reqPer;
+
+                    return reqPer * .25f + m_dependencyOperation.PercentComplete * .75f;
+                }
+            }
             public override TObject ConvertResult(AsyncOperation op)
             {
                 return ((op as UnityWebRequestAsyncOperation).webRequest.downloadHandler as DownloadHandlerTexture).texture as TObject;