+++ package/Runtime/ResourceProviders/Simulation/VirtualBundledAssetProvider.cs
@@ -14,16 +14,23 @@ namespace UnityEngine.ResourceManagement
             where TObject : class
         {
             System.Action<IAsyncOperation<TObject>> onCompleteAction;
+            IAsyncOperation m_dependencyOperation;
+            IAsyncOperation<TObject> m_requestOperation;
+
             public InternalProviderOperation<TObject> Start(IResourceLocation location, IAsyncOperation<IList<object>> loadDependencyOperation)
             {
                 m_result = null;
                 if (onCompleteAction == null)
                     onCompleteAction = OnComplete;
+                m_requestOperation = null;
+                m_dependencyOperation = loadDependencyOperation;
                 loadDependencyOperation.Completed += (obj) => 
                 {
                     VirtualAssetBundle bundle;
                     if (obj.Result != null && obj.Result.Count > 0 && (bundle = obj.Result[0] as VirtualAssetBundle) != null)
-                        bundle.LoadAssetAsync<TObject>(location).Completed += onCompleteAction;
+                    {
+                        (m_requestOperation = bundle.LoadAssetAsync<TObject>(location)).Completed += onCompleteAction;
+                    }
                     else
                         OnComplete();
                 };
@@ -31,6 +38,19 @@ namespace UnityEngine.ResourceManagement
                 return base.Start(location);
             }
 
+            public override float PercentComplete
+            {
+                get
+                {
+                    if (IsDone)
+                        return 1;
+
+                    float reqPer = m_requestOperation == null ? 0 : m_requestOperation.PercentComplete;
+                    if (m_dependencyOperation == null)
+                        return reqPer;
+                    return reqPer * .25f + m_dependencyOperation.PercentComplete * .75f;
+                }
+            }
             public override TObject ConvertResult(AsyncOperation operation) { return null; }
         }