+++ package/Tests/Runtime/ResourceManagerVirtualModeTests.cs
@@ -25,9 +25,8 @@ public class ResourceManagerVirtualModeTests : ResourceManagerBaseTests
         for (int i = 0; i < 10; i++)
         {
             var bundleName = "shared" + i;
-            var isLocal = i % 2 == 0;
             sharedBundles.Add(new VirtualAssetBundle("shared" + i, i % 2 == 0));
-            sharedBundleLocations.Add(new ResourceLocationBase(bundleName, bundleName, isLocal ? typeof(LocalAssetBundleProvider).FullName : typeof(RemoteAssetBundleProvider).FullName));
+            sharedBundleLocations.Add(new ResourceLocationBase(bundleName, bundleName, typeof(AssetBundleProvider).FullName));
         }
         virtualBundleData.AssetBundles.AddRange(sharedBundles);
 
@@ -35,7 +34,7 @@ public class ResourceManagerVirtualModeTests : ResourceManagerBaseTests
         {
             var isLocal = i % 2 == 0;
             var b = new VirtualAssetBundle("bundle" + i, isLocal);
-            var bundleLocation = new ResourceLocationBase(b.Name, b.Name, isLocal ? typeof(LocalAssetBundleProvider).FullName : typeof(RemoteAssetBundleProvider).FullName);
+            var bundleLocation = new ResourceLocationBase(b.Name, b.Name, typeof(AssetBundleProvider).FullName);
             for (int a = 0; a < 10; a++) 
             {
                 var name = b.Name + "_asset" + a;